Fix process exiting prior to sending all data
If relatively large amounts of data was being sent from a child process to the parent (such as a large screenshot), it was possible for the data to not be sent back before the child process executed process.exit because process.send is asynchronous. This was fixed by moving the process.exit into the process.send callback.